+++ src/test/java/com/Java/math/leetcode/LeetCode58.java
@@ -1,0 +1,59 @@
+/**
+ * Copyright (C), 2018-2020, zenki.ai
+ * FileName: LeetCode58
+ * Author:   feiyi
+ * Date:     2020/5/25 9:13 AM
+ * Description:
+ * History:
+ * <author>          <time>          <version>          <desc>
+ * 作者姓名           修改时间           版本号              描述
+ */
+package com.Java.math.leetcode;
+
+import org.junit.Test;
+
+/**
+ * 〈一句话功能简述〉:
+ * 〈最后一个单词的长度〉
+ *
+ * @author feiyi
+ * @create 2020/5/25
+ * @since 1.0.0
+ */
+public class LeetCode58 {
+
+    /**
+     * @Description 给定一个仅包含大小写字母和空格 ' ' 的字符串，返回其最后一个单词的长度。
+     *
+     * 如果不存在最后一个单词，请返回 0 。
+     *
+     * 说明：一个单词是指由字母组成，但不包含任何空格的字符串。
+     *
+     * 示例:
+     *
+     * 输入: "Hello World"
+     * 输出: 5
+     *
+     * 来源：力扣（LeetCode）
+     * 链接：https://leetcode-cn.com/problems/length-of-last-word
+     * 著作权归领扣网络所有。商业转载请联系官方授权，非商业转载请注明出处。
+     * @param
+     * @return: void
+     * @since: 2.0.5
+     * @Author: feiyi
+     * @Date: 2019/7/17 10:53 AM
+     **/
+    @Test
+    public void test4(){
+//        String s = "Hello World";
+        String s = " ";
+        System.out.println(lengthOfLastWord(s));
+    }
+
+    public static int lengthOfLastWord(String s) {
+        if (s == null || s.length() == 0 || s.trim().length() == 0) return 0;
+        s = s.trim();
+        String[] split = s.split(" ");
+        return split[split.length - 1].length();
+    }
+}
